requirements.txt
Flask
Flask-JWT-Extended
Flask-SQLAlchemy
Flask-Migrate
Flask-WTF
email-validator
PyMySQL
openpyxl
bcrypt
python-dotenv
